requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
distro==1.9.0
idna==3.6
iniconfig==2.0.0
langdetect==1.0.9
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.19.1
pypdfium2==4.25.0
pytesseract==0.3.10
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tabula-py==2.9.0
tzdata==2023.4
urllib3==2.1.0
